% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{filterCorrelation}
\alias{filterCorrelation}
\title{Compute the BAHC correlation matrix.}
\usage{
filterCorrelation(x, Nboot = 100)
}
\arguments{
\item{x}{A matrix: $x_{i,f}$ is feature $f$ of object $i$.}

\item{Nboot}{The number of bootstrap copies}
}
\value{
The BAHC-filtered correlation matrix of \code{x}.
}
\description{
Compute the BAHC correlation matrix.
}
\examples{
r=matrix(rnorm(1000),nrow=20)   # 20 objects, 50 features each
Cor_bahc=filterCorrelation(r)
}
